- moved Sorting to FileStorageUtils - Slideshow obeys now the current sorting order
